Use flat imports
This component uses "flat" imports so we don't have too much internal structure exposed in the API. for example you can just `from datamodel import Task, Project` instead of `from datamodel.task import Task; from datamodel.project import Project`

Also don't export strict mode at top level. Should be a rare import.

Remove import sorting hack, as no longer needed
