+++ libs/core/kiln_ai/datamodel/eval_datamodel.py
@@ -1,0 +1,10 @@
+from pydantic import Field
+
+from kiln_ai.datamodel.basemodel import NAME_FIELD, KilnBaseModel
+
+
+class Eval(KilnBaseModel):
+    name: str = NAME_FIELD
+    description: str | None = Field(
+        default=None, description="The description of the eval"
+    )
